Bound catalog shadow read probes
